Update README.md
Problem Statement:

The aim of this project is to build a machine learning model to predict the profit of 50 startups based on their features such as R&D Spend, Marketing Spend, Administration Spend, State, etc. Linear Regression and Ridge Regression are the two models that will be applied on the dataset to determine the best performing model. The presence of outliers in the dataset will be handled using Cook's Distance method. Also, the issue of multicollinearity will be explored and addressed as required.

The objective of this project is to identify the most important features that influence the profit of startups and to create a model that accurately predicts the profit of new startups based on these features. By comparing the performance of the OLS and Ridge Regression models, the best model will be selected to achieve this objective.